+++ docs/changelog.rst
@@ -31,6 +31,8 @@ New Features
 - Multithreading added to :meth:`~gatoh.graphs.graphs.GraphSet.save_graphset`
 - Multiprocessing added to :meth:`~gatoh.model.model.ABModel.calculate_interdependence`
 - Creation of a quickstart guide for the documentation
+- Multiprocessing and multithreading added to :meth:`~gatoh.agents.agents.AgentSet.load_agentset`
+- Multiprocessing and multithreading added to :meth:`~gatoh.graphs.graphs.GraphSet.load_graphset`
 
 Fixes
 -----